+++ framework/Cargo.toml
@@ -19,8 +19,6 @@ aptos-types = { workspace = true, features = ["fuzzing"] }
 bcs = { workspace = true }
 git2 = { workspace = true }
 include_dir = { workspace = true }
-move-core-types = { workspace = true }
-move-command-line-common = { workspace = true }
 once_cell = { workspace = true }
 proptest = { workspace = true, optional = true }
 proptest-derive = { workspace = true, optional = true }
@@ -32,6 +30,10 @@ zapatos-rest-client = { workspace = true }
 zapatos-build-info = { workspace = true }
 zapatos-temppath = { workspace = true }
 
+move-binary-format = { workspace = true }
+move-core-types = { workspace = true }
+move-command-line-common = { workspace = true }
+move-model = { workspace = true }
 
 [build-dependencies]
 zapatos-framework = { workspace = true }
