+++ framework/src/builder/main_generate_proposals.rs
